feat: fix issue

+++ lab report programs/part 3 (56-94)/program56(noRecursive).c
@@ -4,7 +4,7 @@
 
 int main() {
     int num, i;
-    long long factorial = 1;
+    int factorial = 1;
 
     printf("Enter a number: ");
     scanf("%d", &num);
@@ -15,7 +15,7 @@ int main() {
         for (i = 1; i <= num; i++) {
             factorial *= i;
         }
-        printf("Factorial of %d = %lld\n", num, factorial);
+        printf("Factorial of %d = %d\n", num, factorial);
     }
 
     return 0;
