+++ lab report programs/part 3 (56-94)/program57(noRecursive).c
@@ -4,7 +4,7 @@
 
 int main() {
     int base, exponent, i;
-    long long result = 1;
+    int result = 1;
 
     printf("Enter base (b): ");
     scanf("%d", &base);
@@ -18,7 +18,7 @@ int main() {
         for (i = 1; i <= exponent; i++) {
             result *= base;
         }
-        printf("%d^%d = %lld\n", base, exponent, result);
+        printf("%d^%d = %d\n", base, exponent, result);
     }
 
     return 0;
